src/requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
aiohttp-jinja2==1.1.1
astor==0.8.0
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
cmake==3.13.3
cvlib==0.2.1
dlib==19.17.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
httplib2==0.12.3
idna==2.8
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mock==3.0.5
multidict==4.5.2
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
progressbar==2.5
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
trafaret==1.2.0
trafaret-config==2.0.2
urllib3==1.24.1
vk==2.0.2
Werkzeug==0.15.4
yarl==1.3.0
